Abstract the OpenSSL call into the function decrypt_aes_128_ecb
This is the first step, as ideally, the function should not expect
a path as an argument; it should work on the raw bytes. Initial testing,
however, shows that using Matasano.IO.bytes_from_base64/1 and dropping
the `-base64` flag from the openssl command does not produce the same
results.