Fix decrypt_aes_128_ecb to accept raw data rather than a file path
We'll need this for later challenges, and it is a more consistent
interface.

Note that I could not use Elixir.Port's for this purpose, since there is
not way to explicitly send an EOF to STDIN of the process we spawn,
which is unfortunate. See this mailing list thread for a long discussion
and references:

    http://erlang.org/pipermail/erlang-questions/2013-July/074905.html

...so instead, I create a temporary file to have OpenSSL read from. It
works, but is not as robust.